+++ tests/test_instructions.py
@@ -1,0 +1,97 @@
+import pytest
+
+import ida_domain  # isort: skip
+
+from ida_domain.base import InvalidEAError, InvalidParameterError
+
+
+def test_instruction(test_env):
+    db = test_env
+
+    count = 0
+    for _ in db.instructions:
+        count += 1
+    assert count == 197
+
+    instructions = list(db.instructions.get_all())
+    assert len(instructions) == 197
+
+    instruction = db.instructions.get_at(0xD6)
+    assert instruction is not None
+    assert db.instructions.is_valid(instruction)
+    assert db.instructions.get_disassembly(instruction) == 'mov     ax, bx'
+    assert db.instructions.get_operands_count(instruction) == 2
+
+    operands = db.instructions.get_operands(instruction)
+    assert len(operands) == 2
+    assert isinstance(operands[0], ida_domain.operands.RegisterOperand)
+    assert isinstance(operands[1], ida_domain.operands.RegisterOperand)
+
+    assert isinstance(
+        db.instructions.get_operand(instruction, 0), ida_domain.operands.RegisterOperand
+    )
+    assert isinstance(
+        db.instructions.get_operand(instruction, 1), ida_domain.operands.RegisterOperand
+    )
+
+    operands = db.instructions.get_operands(instruction)
+    assert len(operands) == 2
+    assert isinstance(operands[0], ida_domain.operands.RegisterOperand)
+    assert isinstance(operands[1], ida_domain.operands.RegisterOperand)
+
+    instruction = db.instructions.get_previous(0xD6)
+    assert instruction is not None
+    assert instruction.ea == 0xD4
+    assert db.instructions.is_valid(instruction)
+    assert db.instructions.get_disassembly(instruction) == 'mov     eax, ebx'
+    assert db.instructions.get_operands_count(instruction) == 2
+
+    operands = db.instructions.get_operands(instruction)
+    assert len(operands) == 2
+    assert isinstance(operands[0], ida_domain.operands.RegisterOperand)
+    assert isinstance(operands[1], ida_domain.operands.RegisterOperand)
+
+    instructions = list(db.instructions.get_between(0xD0, 0xE0))
+    assert len(instructions) == 7
+
+    instruction = db.instructions.get_at(0xD6)
+    assert instruction is not None
+    mnemonic = db.instructions.get_mnemonic(instruction)
+    assert mnemonic == 'mov'
+
+    # Test get_operand with valid index
+    operand0 = db.instructions.get_operand(instruction, 0)
+    assert operand0 is not None
+    assert isinstance(operand0, ida_domain.operands.RegisterOperand)
+
+    operand1 = db.instructions.get_operand(instruction, 1)
+    assert operand1 is not None
+    assert isinstance(operand1, ida_domain.operands.RegisterOperand)
+
+    # Find a call instruction at 0x262
+    call_insn = db.instructions.get_at(0x262)
+    assert call_insn is not None
+    assert db.instructions.is_call_instruction(call_insn) is True
+    assert db.instructions.is_indirect_jump_or_call(call_insn) is True
+    assert db.instructions.breaks_sequential_flow(call_insn) is False
+
+    # Find a jump instruction at 0x269
+    jmp_insn = db.instructions.get_at(0x269)
+    assert jmp_insn is not None
+    assert db.instructions.is_call_instruction(jmp_insn) is False
+    assert db.instructions.is_indirect_jump_or_call(jmp_insn) is True
+    assert db.instructions.breaks_sequential_flow(jmp_insn) is True
+
+    from ida_domain.base import InvalidEAError, InvalidParameterError
+
+    with pytest.raises(InvalidEAError):
+        list(db.instructions.get_between(0xFFFFFFFF, 0xFFFFFFFF))
+
+    with pytest.raises(InvalidParameterError):
+        list(db.instructions.get_between(0x200, 0x100))
+
+    with pytest.raises(InvalidEAError):
+        db.instructions.get_at(0xFFFFFFFF)
+
+    with pytest.raises(InvalidEAError):
+        db.instructions.get_previous(0xFFFFFFFF)
